models: move enums.file_nodes to models.file_nodes
Related-to: #1030
